Modulo clientes, modulo de Asignación de Unidades

+++ database/migrations/2023_05_10_004921_add_campofaltante_tableunidades_table.php
@@ -14,7 +14,7 @@ class AddCampofaltanteTableunidadesTable extends Migration
     public function up()
     {
         Schema::table('unidades', function (Blueprint $table) {
-            $table->double('valorAdquisicionUnidad',40,2)->default(null)->after('valorActualunidad');
+            $table->decimal('valorAdquisicionUnidad',15,2)->default(null)->after('valorActualunidad');
             $table->string('noserie',100)->nullable()->after('valorAdquisicionUnidad');
         });
     }
